moves deployment pills into deploymentProvider
