+++ shared/agent/src/providers/newrelic/goldenSignals/goldenSignalsProvider.ts
@@ -18,7 +18,6 @@ import {
 	MetricTimesliceNameMapping,
 	NRErrorResponse,
 	ObservabilityRepo,
-	Pills,
 } from "@codestream/protocols/agent";
 import { CSMe } from "@codestream/protocols/api";
 import { uniqBy as _uniqBy } from "lodash";
@@ -27,18 +26,20 @@ import { Logger } from "../../../logger";
 import { log } from "../../../system/decorators/log";
 import { lsp, lspHandler } from "../../../system/decorators/lsp";
 import { ContextLogger } from "../../contextLogger";
+import { DeploymentsProvider } from "../deployments/deploymentsProvider";
 import { NewRelicGraphqlClient, escapeNrql } from "../newRelicGraphqlClient";
 import { NrApiConfig } from "../nrApiConfig";
+import { NrqlQueryBuilder } from "../nrql/nrqlQueryBuilder";
 import { ReposProvider } from "../repos/reposProvider";
 import { mapNRErrorResponse, parseId, toFixedNoRounding } from "../utils";
-import { NrqlQueryBuilder } from "../nrql/nrqlQueryBuilder";
 
 @lsp
 export class GoldenSignalsProvider {
 	constructor(
 		private graphqlClient: NewRelicGraphqlClient,
 		private reposProvider: ReposProvider,
-		private nrApiConfig: NrApiConfig
+		private nrApiConfig: NrApiConfig,
+		private deploymentsProvider: DeploymentsProvider
 	) {}
 
 	@lspHandler(GetServiceLevelTelemetryRequestType)
@@ -92,148 +93,6 @@ export class GoldenSignalsProvider {
 			};
 			return response;
 		}
-		return undefined;
-	}
-
-	async getPillsData(entityGuid: string, accountId?: number): Promise<Pills | undefined> {
-		if (!entityGuid && !accountId) {
-			return undefined;
-		}
-
-		try {
-			const countQuery = [
-				"SELECT",
-				"latest(deploymentId) AS 'deploymentId',",
-				"latest(timestamp) AS 'timestamp'",
-				"FROM Deployment",
-				`WHERE entity.guid = '${entityGuid}'`,
-				"SINCE 13 months ago",
-			].join(" ");
-
-			const countResponse = await this.graphqlClient.query<{
-				actor: {
-					account: {
-						nrql: {
-							results: {
-								deploymentId: string;
-								timestamp: number;
-							}[];
-						};
-					};
-					entity: {
-						permalink: string;
-					};
-				};
-			}>(
-				`query fetchErrorRate($accountId:Int!, $entityGuid:EntityGuid!) {
-				actor {
-					account(id: $accountId) {
-						nrql(
-							options: { eventNamespaces: "Marker" }
-							query: "${countQuery}"
-						) { nrql results }
-					}
-					entity(guid: $entityGuid) {
-						permalink
-					}
-				}
-			}`,
-				{
-					accountId: accountId,
-					entityGuid: entityGuid,
-				}
-			);
-
-			const countResults = countResponse.actor.account.nrql?.results[0];
-
-			const { deploymentId, timestamp } = countResults;
-
-			if (!deploymentId || !timestamp) return undefined;
-
-			const deployListUrl = this.deltaUrl(entityGuid);
-
-			const threeHoursLater = timestamp + 10800000;
-
-			const comparisonQuery = [
-				"FROM Metric",
-				"SELECT",
-				"average(newrelic.goldenmetrics.apm.application.responseTimeMs) AS 'responseTimeMs',",
-				"average(newrelic.goldenmetrics.apm.application.errorRate) AS 'errorRate'",
-				`WHERE entity.guid = '${entityGuid}'`,
-				`SINCE ${timestamp} UNTIL ${threeHoursLater}`,
-				`COMPARE WITH 3 hours ago`,
-			].join(" ");
-
-			const comparisonQueryData = await this.graphqlClient.query<{
-				actor: {
-					account: {
-						nrql: {
-							results: {
-								beginTimeSeconds: number;
-								endTimeSeconds: number;
-								responseTimeMs: number;
-								errorRate: number;
-							}[];
-						};
-					};
-				};
-			}>(
-				`query fetchErrorRate($accountId:Int!) {
-					actor {
-						account(id: $accountId) {
-							nrql(
-								
-								query: "${comparisonQuery}"
-							) { nrql results }
-						}
-					}
-				}`,
-				{
-					accountId: accountId,
-				}
-			);
-
-			const currentMetrics = comparisonQueryData.actor.account.nrql?.results[0];
-			const previousMetrics = comparisonQueryData.actor.account.nrql?.results[1];
-
-			if (!currentMetrics || !previousMetrics) return undefined;
-
-			const errorRatePercentageChange =
-				previousMetrics.errorRate > 0
-					? Math.round(
-							((currentMetrics.errorRate - previousMetrics.errorRate) / previousMetrics.errorRate) *
-								100
-					  )
-					: undefined;
-			const responseTimePercentageChange =
-				previousMetrics.responseTimeMs > 0
-					? Math.round(
-							((currentMetrics.responseTimeMs - previousMetrics.responseTimeMs) /
-								previousMetrics.responseTimeMs) *
-								100
-					  )
-					: undefined;
-
-			return {
-				errorRateData: {
-					percentChange:
-						errorRatePercentageChange && errorRatePercentageChange >= 0
-							? errorRatePercentageChange
-							: undefined,
-					permalinkUrl: deployListUrl,
-				},
-				responseTimeData: {
-					percentChange:
-						responseTimePercentageChange && responseTimePercentageChange >= 0
-							? responseTimePercentageChange
-							: undefined,
-					permalinkUrl: deployListUrl,
-				},
-			};
-		} catch (err) {
-			Logger.error(err, entityGuid);
-		}
-
 		return undefined;
 	}
 
@@ -339,23 +198,6 @@ export class GoldenSignalsProvider {
 		}
 
 		return `${bUrl}/accounts/${accountId}/issues/${issueId}?notifierType=codestream`;
-	}
-
-	private deltaUrl(entityGuid: string) {
-		let bUrl = "";
-
-		if (this.nrApiConfig.productUrl.includes("staging")) {
-			// Staging
-			bUrl = "https://staging-one.newrelic.com";
-		} else if (this.nrApiConfig.productUrl.includes("eu")) {
-			// EU
-			bUrl = "https://one.eu.newrelic.com";
-		} else {
-			// Prod:
-			bUrl = "https://one.newrelic.com";
-		}
-
-		return `${bUrl}/nr1-core/deployment-markers/list/${entityGuid}`;
 	}
 
 	// Map array of objects based on order of array of strings
@@ -702,7 +544,7 @@ export class GoldenSignalsProvider {
 				};
 			});
 
-			const pillsData = await this.getPillsData(entityGuid, accountId);
+			const pillsData = await this.deploymentsProvider.getDeploymentDiff(entityGuid, accountId);
 
 			return {
 				lastUpdated: new Date().toLocaleString(),
